add confirm to deleted and modified localstorage for session storage

+++ FrontEnd/assets/scripts/delete-works.js
@@ -8,9 +8,13 @@ export async function deleteWorks(e) {
     const works = input.parentNode;
     const worksId = works.dataset.id;
 
-    deleteWorksData(worksId);
-    const getWorks = await getWorksData();
-    createdGallery(getWorks, ".modal-gallery", true);
-    createdGallery(getWorks, ".gallery", false);
-    getBtTrash();
+    let validDelete = confirm("Êtes vous sur de vouloir supprimer le projet ?");
+    if (validDelete) {
+        deleteWorksData(worksId);
+        const getWorks = await getWorksData();
+        createdGallery(getWorks, ".modal-gallery", true);
+        createdGallery(getWorks, ".gallery", false);
+        getBtTrash();
+    }
+
 }
